+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog ##
 #### dev ####
 * Add check to Prefix for ByRef not being array type
+* Fix ByRef not working on array types
 
 #### v3.10.1 ####
 * Fix crash in incomplete last-update code